Move FooByCall* to sub-package
