+++ src/main/c++/FooByCallFinal.cpp
@@ -25,7 +25,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 #include <jni.h>
-#include "com_evolvedbinary_jnibench_common_FooByCallFinal.h"
+#include "com_evolvedbinary_jnibench_common_call_FooByCallFinal.h"
 #include "Foo.h"
 
 /*
@@ -33,7 +33,7 @@
  * Method:    newFoo
  * Signature: ()J
  */
-jlong Java_com_evolvedbinary_jnibench_common_FooByCallFinal_newFoo(JNIEnv* env, jobject jobj) {
+jlong Java_com_evolvedbinary_jnibench_common_call_FooByCallFinal_newFoo(JNIEnv* env, jobject jobj) {
   jnibench::Foo* foo = new jnibench::Foo();
   return reinterpret_cast<jlong>(foo);
 }
@@ -43,6 +43,6 @@ jlong Java_com_evolvedbinary_jnibench_common_FooByCallFinal_newFoo(JNIEnv* env, 
  * Method:    disposeInternal
  * Signature: (J)V
  */
-void Java_com_evolvedbinary_jnibench_common_FooByCallFinal_disposeInternal(JNIEnv* env, jobject jobj, jlong handle) {
+void Java_com_evolvedbinary_jnibench_common_call_FooByCallFinal_disposeInternal(JNIEnv* env, jobject jobj, jlong handle) {
     delete reinterpret_cast<jnibench::Foo*>(handle);
 }
