+++ src/main/c++/FooByCallInvokeFinal.cpp
@@ -25,7 +25,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 #include <jni.h>
-#include "com_evolvedbinary_jnibench_common_FooByCallInvokeFinal.h"
+#include "com_evolvedbinary_jnibench_common_call_FooByCallInvokeFinal.h"
 #include "Foo.h"
 #include "Portal.h"
 
@@ -34,7 +34,7 @@
  * Method:    newFoo
  * Signature: ()J
  */
-void Java_com_evolvedbinary_jnibench_common_FooByCallInvokeFinal_newFoo(JNIEnv* env, jobject jobj) {
+void Java_com_evolvedbinary_jnibench_common_call_FooByCallInvokeFinal_newFoo(JNIEnv* env, jobject jobj) {
   jnibench::Foo* foo = new jnibench::Foo();
 
   //set the _nativeHandle in Java
@@ -46,6 +46,6 @@ void Java_com_evolvedbinary_jnibench_common_FooByCallInvokeFinal_newFoo(JNIEnv* 
  * Method:    disposeInternal
  * Signature: (J)V
  */
-void Java_com_evolvedbinary_jnibench_common_FooByCallInvokeFinal_disposeInternal(JNIEnv* env, jobject jobj, jlong handle) {
+void Java_com_evolvedbinary_jnibench_common_call_FooByCallInvokeFinal_disposeInternal(JNIEnv* env, jobject jobj, jlong handle) {
     delete reinterpret_cast<jnibench::Foo*>(handle);
 }
