+++ src/main/java/com/evolvedbinary/jnibench/common/call/FooByCallInvoke.java
@@ -24,15 +24,18 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.evolvedbinary.jnibench.common;
+package com.evolvedbinary.jnibench.common.call;
+
+import com.evolvedbinary.jnibench.common.NativeBackedObject;
 
 /**
- * Similar to {@link FooByCallInvoke} but this class is marked final.
+ * Follows <i>9.2.7 Pattern 4: Call-Invoke</i> from Java Platform Performance by Steve Wilson
+ * for setting up the handle to the native object.
  *
  * @author <a href="mailto:adam@evolvedbinary.com">Adam Retter</a>
  */
-public final class FooByCallInvokeFinal extends NativeBackedObject {
-    public FooByCallInvokeFinal() {
+public class FooByCallInvoke extends NativeBackedObject {
+    public FooByCallInvoke() {
         super();
         newFoo();   //the native method, will find _nativeHandle from the class and set it directly
     }
